# docs/design/openapi.yaml
openapi: 3.1.0
info:
  title: RMS Implementation Architecture Baseline 2.0 — OpenAPI draft
  version: 2.0.0-draft
  description: Design only; NOT IMPLEMENTED. 业务来源唯一冻结Baseline1.3。服务器根/api/v1。全部保护操作检查session、启用状态、实际角色并集；自我评审等限制独立执行。请求只白名单，历史不可变，无generic
    status API。SQL BIGINT IDs/lockVersion采用十进制字符串。两个/auth技术GET为实现派生，不新增FR。数据库generated活动键不公开。密码原值与1..72 UTF-8 bytes技术政策已冻结；非admin assignee目录待决。
servers:
- url: /api/v1
security:
- cookieSession: []
...
paths:
  /auth/login:
    post:
      operationId: login
      tags:
      - auth
      summary: 用户名或邮箱登录；禁用403，未知或错误凭证401，成功轮换会话ID
      description: 用户名或邮箱登录；禁用403，未知或错误凭证401，成功轮换会话ID
      x-module: auth
      x-fr:
      - FR-AUTH-001
      - FR-AUTH-003
      x-br:
      - BR-AUTH-001
      - BR-AUTH-002
      - BR-AUTH-003
      - BR-AUTH-004
      x-uc:
      - UC-01
      x-ac:
      - AC-01
      x-required-roles: []
      x-transaction: SECURITY_SESSION_LIFECYCLE
      responses:
        '200':
          description: 成功；具体业务含义见操作说明
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionResponse'
        '400':
          description: 字段/完整性错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          description: 未认证/凭证错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: 权限/启用状态/自我评审/CSRF拒绝
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: 目标不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '409':
          description: 修订/状态/唯一/基准/图锁冲突
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '500':
          description: 内部错误；业务写全回滚
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
      security: []
      parameters:
      - $ref: '#/components/parameters/CsrfHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
  /auth/logout:
    post:
      operationId: logout
      tags:
      - auth
      summary: 使当前已认证会话失效
      description: 使当前已认证会话失效。仅启用账号；服务端重新检查角色并集及独立业务限制。
      x-module: auth
      x-fr:
      - FR-AUTH-002
      x-br:
      - BR-AUTH-001
      - BR-AUTH-002
      - BR-AUTH-003
      - BR-AUTH-004
      x-uc:
      - UC-01
      x-ac:
      - AC-01
      x-required-roles: &id001
      - ADMIN
      - REQUIREMENT_ENGINEER
      - REVIEWER
      - PROJECT_MEMBER
      - VIEWER
      x-transaction: SECURITY_SESSION_LIFECYCLE
      responses:
        '204':
          description: 成功；具体业务含义见操作说明
        '400':
          description: 字段/完整性错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          description: 未认证/凭证错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: 权限/启用状态/自我评审/CSRF拒绝
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: 目标不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '409':
          description: 修订/状态/唯一/基准/图锁冲突
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '500':
          description: 内部错误；业务写全回滚
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
      parameters:
      - $ref: '#/components/parameters/CsrfHeader'
  /auth/csrf:
    get:
      operationId: getCsrf
      tags:
      - auth
      summary: 实现派生：提供CSRF令牌，登录和退出后重新获取；不是bootstrap业务接口
      description: 实现派生：提供CSRF令牌，登录和退出后重新获取；不是bootstrap业务接口
      x-module: auth
      x-fr:
      - FR-AUTH-001
      - FR-AUTH-003
      x-br:
      - BR-AUTH-001
      - BR-AUTH-002
      - BR-AUTH-003
      - BR-AUTH-004
      x-uc:
      - UC-01
      x-ac:
      - AC-01
      x-required-roles: []
      x-transaction: SECURITY_SESSION_READ
      responses:
        '200':
          description: 成功；具体业务含义见操作说明
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CsrfResponse'
        '400':
          description: 字段/完整性错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          description: 未认证/凭证错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: 权限/启用状态/自我评审/CSRF拒绝
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: 目标不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '409':
          description: 修订/状态/唯一/基准/图锁冲突
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '500':
          description: 内部错误；业务写全回滚
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
      security: []
  /auth/session:
    get:
      operationId: getSession
      tags:
      - auth
      summary: 实现派生：读取当前可信会话身份和实际角色
      description: 实现派生：读取当前可信会话身份和实际角色。仅启用账号；服务端重新检查角色并集及独立业务限制。
      x-module: auth
      x-fr:
      - FR-AUTH-003
      - FR-USER-006
      x-br:
      - BR-AUTH-001
      - BR-AUTH-002
      - BR-AUTH-003
      - BR-AUTH-004
      - BR-CR-019
      - BR-REV-004
      x-uc:
      - UC-01
      - UC-02
      x-ac:
      - AC-01
      x-required-roles: *id001
      x-transaction: SECURITY_SESSION_READ
      responses:
        '200':
          description: 成功；具体业务含义见操作说明
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionResponse'
        '400':
          description: 字段/完整性错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          description: 未认证/凭证错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: 权限/启用状态/自我评审/CSRF拒绝
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: 目标不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '409':
          description: 修订/状态/唯一/基准/图锁冲突
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '500':
          description: 内部错误；业务写全回滚
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
...
  /requirements/{id}/history:
    get:
      operationId: listRequirementHistory
      tags:
      - audit
      summary: 当前需求业务审计只读，版本与评审历史通过独立只读入口查看
      description: 当前需求业务审计只读，版本与评审历史通过独立只读入口查看。仅启用账号；服务端重新检查角色并集及独立业务限制。
      x-module: audit
      x-fr:
      - FR-AUDIT-004
      x-br:
      - BR-AUDIT-002
      - BR-AUDIT-005
      - BR-AUTH-001
      - BR-AUTH-002
      - BR-AUTH-003
      - BR-AUTH-004
      x-uc:
      - UC-08
      x-ac:
      - AC-10
      x-required-roles: *id001
      x-transaction: READ_ONLY
      responses:
        '200':
          description: 成功；具体业务含义见操作说明
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuditResponsePage'
        '400':
          description: 字段/完整性错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          description: 未认证/凭证错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: 权限/启用状态/自我评审/CSRF拒绝
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: 目标不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '409':
          description: 修订/状态/唯一/基准/图锁冲突
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '500':
          description: 内部错误；业务写全回滚
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
      parameters:
      - name: id
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/Id'
      - $ref: '#/components/parameters/Page'
      - $ref: '#/components/parameters/Size'
      - &id006
        name: from
        in: query
        required: false
        schema:
          type: string
          format: date-time
        description: UTC过滤起点
      - &id007
        name: to
        in: query
        required: false
        schema:
          type: string
          format: date-time
        description: UTC过滤终点
      - &id008
        name: action
        in: query
        required: false
        schema:
          type: string
          maxLength: 64
        description: action
  /audit-events:
    get:
      operationId: listAuditEvents
      tags:
      - audit
      summary: 全系统业务审计只读ADMIN-only，安全字段白名单，无密码或凭证
      description: 全系统业务审计只读ADMIN-only，安全字段白名单，无密码或凭证。仅启用账号；服务端重新检查角色并集及独立业务限制。
      x-module: audit
      x-fr:
      - FR-AUDIT-004
      x-br:
      - BR-AUDIT-002
      - BR-AUDIT-005
      - BR-AUTH-001
      - BR-AUTH-002
      - BR-AUTH-003
      - BR-AUTH-004
      x-uc:
      - UC-08
      x-ac:
      - AC-10
      x-required-roles: *id002
      x-transaction: READ_ONLY
      responses:
        '200':
          description: 成功；具体业务含义见操作说明
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuditResponsePage'
        '400':
          description: 字段/完整性错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          description: 未认证/凭证错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: 权限/启用状态/自我评审/CSRF拒绝
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: 目标不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '409':
          description: 修订/状态/唯一/基准/图锁冲突
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '500':
          description: 内部错误；业务写全回滚
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
      parameters:
      - $ref: '#/components/parameters/Page'
      - $ref: '#/components/parameters/Size'
      - *id006
      - *id007
      - *id008
components:
  securitySchemes:
    cookieSession:
      type: apiKey
      in: cookie
      name: JSESSIONID
      description: HttpOnly Secure SameSite=Lax，同源session；所有unsafe请求另需X-CSRF-TOKEN，包括login/logout。
  parameters:
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 0
        default: 0
    Size:
      name: size
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    CsrfHeader:
      name: X-CSRF-TOKEN
      in: header
      required: true
      schema:
        type: string
        minLength: 1
      description: GET /auth/csrf取得；登录/退出后刷新；不是actor身份。
  schemas:
    Id:
      type: string
      pattern: ^[1-9][0-9]*$
      maxLength: 19
      description: SQL signed BIGINT正整数十进制字符串；服务端再校验<=9223372036854775807，避免JavaScript Number精度丢失。
    LockVersion:
      type: string
      pattern: ^(0|[1-9][0-9]*)$
      maxLength: 19
      description: 原有signed BIGINT lockVersion非负十进制字符串；服务端校验<=9223372036854775807、比较并由@Version更新。
    RoleCode:
      type: string
      enum:
      - ADMIN
      - REQUIREMENT_ENGINEER
      - REVIEWER
      - PROJECT_MEMBER
      - VIEWER
...
    UserSummary:
      type: object
      additionalProperties: false
      properties:
        userId:
          $ref: '#/components/schemas/Id'
        displayName:
          type: string
          maxLength: 100
        accountStatus:
          type: string
          enum:
          - ENABLED
          - DISABLED
      required:
      - userId
      - displayName
      - accountStatus
...
    AuditResponse:
      type: object
      additionalProperties: false
      properties:
        auditId:
          type: string
          pattern: ^[1-9][0-9]*$
          maxLength: 19
          description: SQL signed BIGINT正整数十进制字符串；服务端再校验<=9223372036854775807，避免JavaScript Number精度丢失。
          readOnly: true
        actorId:
          type: string
          pattern: ^[1-9][0-9]*$
          maxLength: 19
          description: SQL signed BIGINT正整数十进制字符串；服务端再校验<=9223372036854775807，避免JavaScript Number精度丢失。
          readOnly: true
        requirementId:
          type:
          - string
          - 'null'
          pattern: ^[1-9][0-9]*$
          maxLength: 19
          description: SQL signed BIGINT正整数十进制字符串；服务端再校验<=9223372036854775807，避免JavaScript Number精度丢失。
          readOnly: true
        targetType:
          type: string
          maxLength: 32
          enum:
          - USER
          - ROLE_ASSIGNMENT
          - REQUIREMENT
          - REVIEW
          - VERSION
          - RELATION
          - CHANGE_REQUEST
          - CHANGE_REVIEW
          - TAG_ASSIGNMENT
          - TAG
          - COMMENT
          readOnly: true
        targetId:
          type: string
          maxLength: 128
          readOnly: true
        action:
          type: string
          maxLength: 64
          readOnly: true
        outcome:
          type: string
          maxLength: 32
          enum:
          - SUCCESS
          - DENIED
          - FAILED
          readOnly: true
        beforeData:
          type:
          - object
          - 'null'
          additionalProperties: true
          description: 安全字段白名单审计投影，严禁凭证/密码/哈希。
          readOnly: true
        afterData:
          type:
          - object
          - 'null'
          additionalProperties: true
          description: 安全字段白名单审计投影，严禁凭证/密码/哈希。
          readOnly: true
        occurredAt:
          type: string
          format: date-time
          pattern: .*Z$
          description: 服务端UTC，ISO8601 Z；SQL DATETIME(6)，微秒精度。
          readOnly: true
      required:
      - auditId
      - actorId
      - requirementId
      - targetType
      - targetId
      - action
      - outcome
      - beforeData
      - afterData
      - occurredAt
      description: 审计只读，安全投影，排除任何密码/哈希/会话或CSRF凭证。
    SessionResponse:
      type: object
      additionalProperties: false
      properties:
        user:
          $ref: '#/components/schemas/UserSummary'
        roles:
          type: array
          items:
            $ref: '#/components/schemas/RoleCode'
          minItems: 1
          uniqueItems: true
      required:
      - user
      - roles
    CsrfResponse:
      type: object
      additionalProperties: false
      properties:
        headerName:
          type: string
          const: X-CSRF-TOKEN
        token:
          type: string
          minLength: 1
      required:
      - headerName
      - token
    LoginRequest:
      type: object
      additionalProperties: false
      properties:
        login:
          type: string
          maxLength: 254
          minLength: 1
          pattern: .*\S.*
        password:
          type: string
          minLength: 1
          pattern: .*\S.*
          writeOnly: true
          x-rms-max-utf8-bytes: 72
          description: 原始Java String；不trim、不做NFC/NFKC或其他Unicode normalization。满足既有non-blank contract且按原值UTF-8编码的实际长度为1..72 bytes；超过72 bytes以400 INVALID_INPUT拒绝，不静默截断。与创建用户initialPassword使用同一服务端byte-limit validator；不得回显、记录或持久化原密码。maxLength不代表字节限制；扩展属性须由服务端显式实现。
      required:
      - login
      - password
...
    ApiError:
      type: object
      additionalProperties: false
      properties:
        code:
          type: string
          enum:
          - INVALID_INPUT
          - INCOMPLETE_CONTENT
          - UNAUTHENTICATED
          - INVALID_CREDENTIALS
          - FORBIDDEN
          - ACCOUNT_DISABLED
          - SELF_REVIEW
          - NOT_AUTHOR
          - NOT_FOUND
          - LOCK_VERSION_CONFLICT
          - STATE_CONFLICT
          - BASE_VERSION_STALE
          - DUPLICATE
          - ACTIVE_CHANGE_EXISTS
          - ACTIVE_REVIEW_EXISTS
          - DEPENDENCY_CYCLE
          - GRAPH_BUSY
          - INTERNAL_ERROR
        message:
          type: string
        correlationId:
          type: string
        currentLockVersion:
          $ref: '#/components/schemas/LockVersion'
      required:
      - code
      - message
      - correlationId
      description: 安全错误，无SQL/stack/密码；currentLockVersion仅有明确受权资源的修订冲突才提供。
...
    AuditResponsePage:
      type: object
      additionalProperties: false
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/AuditResponse'
        page:
          type: integer
          minimum: 0
        size:
          type: integer
          minimum: 1
          maximum: 100
        totalElements:
          type: integer
          minimum: 0
      required:
      - items
      - page
      - size
      - totalElements